docs: added inline-comment-doc
